+++ src/Filters/RbacServiceFilters.php
@@ -7,6 +7,7 @@ use Bayfront\Bones\Application\Services\Filters\FilterSubscription;
 use Bayfront\Bones\Interfaces\FilterSubscriberInterface;
 use Bayfront\BonesService\Rbac\Migrations\CreateRbacServiceSchema;
 use Bayfront\BonesService\Rbac\Migrations\UpdateRbacServiceSchema_v1_1;
+use Bayfront\BonesService\Rbac\Migrations\UpdateRbacServiceSchema_v1_3;
 use Bayfront\BonesService\Rbac\RbacService;
 
 class RbacServiceFilters extends FilterSubscriber implements FilterSubscriberInterface
@@ -33,7 +34,8 @@ class RbacServiceFilters extends FilterSubscriber implements FilterSubscriberInt
     {
         return array_merge($array, [
             new CreateRbacServiceSchema($this->rbacService),
-            new UpdateRbacServiceSchema_v1_1($this->rbacService)
+            new UpdateRbacServiceSchema_v1_1($this->rbacService),
+            new UpdateRbacServiceSchema_v1_3($this->rbacService)
         ]);
     }